Clean up custom-aggrid-header
